enh/absolute-import: Use absolute import for all components and models except for sibling imports and Rename all component scss files into styles.module.scss
