+++ src/App.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import Map from './components/Map';
-import HeadNav from './components/HeadNav';
-import RecordButton from './components/RecordButton';
+import Map from 'components/Map';
+import HeadNav from 'components/HeadNav';
+import RecordButton from 'components/RecordButton';
 import classes from './App.module.scss';
 
 const App: React.FC = () => {
