+++ src/components/RecordButton/styles.module.scss
